abc137/src/test/java/abc137/c/MainTest.java
package abc137.c;

import org.junit.Test;

import java.io.ByteArrayInputStream;
import java.io.ByteArrayOutputStream;
import java.io.PrintStream;

import static org.hamcrest.CoreMatchers.is;
import static org.hamcrest.MatcherAssert.assertThat;

public class MainTest {

    @Test
    public void test1() {
        String input = "3\n" +
                "acornistnt\n" +
                "peanutbomb\n" +
                "constraint";
        String expected = "1";

        assertThat(execute(input + "\n"), is(expected + "\n"));
    }

    @Test
    public void test2() {
        String input = "2\n" +
                "oneplustwo\n" +
                "ninemodsix";
        String expected = "0";

        assertThat(execute(input + "\n"), is(expected + "\n"));
    }

    @Test
    public void test3() {
        String input = "5\n" +
                "abaaaaaaaa\n" +
                "oneplustwo\n" +
                "aaaaaaaaba\n" +
                "twoplusone\n" +
                "aaaabaaaaa";
        String expected = "4";

        assertThat(execute(input + "\n"), is(expected + "\n"));
    }

    private String execute(String input) {
        ByteArrayOutputStream os = new ByteArrayOutputStream();
        Main.solve(
                new ByteArrayInputStream(input.getBytes()), new PrintStream(os)
        );
        return os.toString().replace(System.lineSeparator(), "\n");
    }
}
